加入圖片 完成進階消除 進行目標設計 Ver0.5

+++ candy_control.h
@@ -21,7 +21,7 @@ public:
     int add(int style);
     void show_score(QPoint pos, int num);
     int score,scoring;
-    int combo;
+    double combo;
     QLabel *label;
     QWidget *par;
 };
